test(oldms): split oldms_test into per-format source files
Move the doc/ppt/xls tests into doc_test.cpp, ppt_test.cpp and
xls_test.cpp; the shared collect_text helper moves to
oldms_test_util.hpp (mirroring pdf_test_file_builder). No test changes.
