+++ test/CMakeLists.txt
@@ -35,7 +35,9 @@ add_executable(odr_test
         "src/internal/csv/csv_file_test.cpp"
         "src/internal/csv/csv_test.cpp"
 
-        "src/internal/oldms/oldms_test.cpp"
+        "src/internal/oldms/doc_test.cpp"
+        "src/internal/oldms/ppt_test.cpp"
+        "src/internal/oldms/xls_test.cpp"
 
         "src/internal/ooxml/ooxml_crypto_test.cpp"
