+++ test/src/internal/oldms/ppt_test.cpp
@@ -1,0 +1,96 @@
+#include <test_util.hpp>
+
+#include <gtest/gtest.h>
+
+#include <odr/document.hpp>
+#include <odr/document_element.hpp>
+#include <odr/file.hpp>
+#include <odr/odr.hpp>
+
+#include <internal/oldms/oldms_test_util.hpp>
+
+#include <memory>
+#include <string>
+#include <vector>
+
+using namespace odr;
+using namespace odr::test;
+using odr::test::oldms::collect_text;
+
+TEST(OldMs, ppt_empty) {
+  const std::unique_ptr logger =
+      Logger::create_stdio("odr-test", LogLevel::verbose);
+
+  const DocumentFile document_file(
+      TestData::test_file_path("odr-public/ppt/empty.ppt"), *logger);
+
+  EXPECT_EQ(document_file.file_type(),
+            FileType::legacy_powerpoint_presentation);
+
+  const Document document = document_file.document();
+  EXPECT_EQ(document.document_type(), DocumentType::presentation);
+
+  std::size_t slide_count = 0;
+  for (const Element slide : document.root_element().children()) {
+    EXPECT_EQ(slide.type(), ElementType::slide);
+    ++slide_count;
+  }
+  EXPECT_EQ(slide_count, 1);
+}
+
+TEST(OldMs, ppt_style_various) {
+  const std::unique_ptr logger =
+      Logger::create_stdio("odr-test", LogLevel::verbose);
+
+  const DocumentFile document_file(
+      TestData::test_file_path("odr-public/ppt/style-various-1.ppt"), *logger);
+
+  EXPECT_EQ(document_file.file_type(),
+            FileType::legacy_powerpoint_presentation);
+
+  const Document document = document_file.document();
+  EXPECT_EQ(document.document_type(), DocumentType::presentation);
+
+  // The text boxes (frames) of each slide, in shape order.
+  std::vector<std::vector<Element>> slides;
+  for (const Element slide : document.root_element().children()) {
+    EXPECT_EQ(slide.type(), ElementType::slide);
+    std::vector<Element> frames;
+    for (const Element child : slide.children()) {
+      ASSERT_EQ(child.type(), ElementType::frame);
+      frames.push_back(child);
+    }
+    ASSERT_FALSE(frames.empty()); // every slide has at least its title box
+    slides.push_back(std::move(frames));
+  }
+  ASSERT_EQ(slides.size(), 8);
+
+  // Every frame is positioned (anchored to the page, all four measures
+  // present).
+  std::size_t total_frames = 0;
+  for (const std::vector<Element> &frames : slides) {
+    total_frames += frames.size();
+    for (const Element &element : frames) {
+      const Frame frame = element.as_frame();
+      EXPECT_EQ(frame.anchor_type(), AnchorType::at_page);
+      EXPECT_TRUE(frame.x().has_value());
+      EXPECT_TRUE(frame.y().has_value());
+      EXPECT_TRUE(frame.width().has_value());
+      EXPECT_TRUE(frame.height().has_value());
+    }
+  }
+  EXPECT_EQ(total_frames, 12);
+
+  // The first frame of each slide is its title "titleN…", in order.
+  for (std::size_t i = 0; i < slides.size(); ++i) {
+    const std::string title = collect_text(slides[i].front());
+    EXPECT_EQ(title.rfind("title" + std::to_string(i + 1), 0), 0u)
+        << "slide " << i << " title: " << title;
+  }
+
+  // Slide 0 is a title + subtitle box, at different vertical positions.
+  ASSERT_EQ(slides[0].size(), 2);
+  EXPECT_EQ(collect_text(slides[0][0]), "title1");
+  EXPECT_EQ(collect_text(slides[0][1]), "subtitle");
+  EXPECT_NE(slides[0][0].as_frame().y(), slides[0][1].as_frame().y());
+}
